Phase 4 Wave A: move party mode into surfaces/party/
Second git-mv batch of the core/+surfaces/ scaffolding. PartyPage,
PartyDiagram, and the party-specific half of mp/ (ControllerView,
PartyStage, QrCode, mpApi, roomClient, voicePhrases) move to
apps/web/src/surfaces/party/.

mp/protocol.ts stays put: an exhaustive grep found it imported by 14
files outside party (zone.ts, world.ts, presence.ts, sensing.ts,
voiceMeter.ts, gamepadSpike.ts, gameSave.ts and their tests) — it's a
shared realtime primitive, not party-specific, so forcing it into this
surface would be wrong. Same call as leaving NavHeader's admin import
alone in the previous batch: no premature core/ extraction without a
boundary rule to enforce it.

App.tsx's PartyPage lazy import and its static ControllerView/
PartyStage/mpApi imports repointed at the new paths; nothing else in
the tree referenced the old locations.

Verified: tsc clean, full apps/web suite (202 files / 1775 tests)
green, repo lint exits 0, production build succeeds, Playwright check
of / -> Party nav renders PartyPage's real copy correctly.
