+++ apps/web/src/surfaces/party/PartyDiagram.tsx
