+++ apps/web/src/surfaces/party/QrCode.tsx
