+++ eslint-rules/comment-prose-baseline.json
@@ -584,6 +584,12 @@
     "apps/web/src/SubmissionStatusView.tsx": 4172,
     "apps/web/src/SuggestionsPanel.test.tsx": 28,
     "apps/web/src/SuggestionsPanel.tsx": 113,
+    "apps/web/src/surfaces/party/ControllerView.tsx": 127,
+    "apps/web/src/surfaces/party/mpApi.ts": 23,
+    "apps/web/src/surfaces/party/PartyStage.tsx": 171,
+    "apps/web/src/surfaces/party/QrCode.tsx": 38,
+    "apps/web/src/surfaces/party/roomClient.ts": 92,
+    "apps/web/src/surfaces/party/voicePhrases.ts": 13,
     "apps/web/src/telemetry.test.ts": 201,
     "apps/web/src/telemetry.ts": 968,
     "apps/web/src/TelemetryCharts.test.tsx": 26,
