+++ eslint-rules/module-size-baseline.json
@@ -834,6 +834,17 @@
     "apps/web/src/SubmissionStatusView.tsx": 2649,
     "apps/web/src/SuggestionsPanel.test.tsx": 115,
     "apps/web/src/SuggestionsPanel.tsx": 146,
+    "apps/web/src/surfaces/party/ControllerView.test.tsx": 147,
+    "apps/web/src/surfaces/party/ControllerView.tsx": 328,
+    "apps/web/src/surfaces/party/mpApi.ts": 41,
+    "apps/web/src/surfaces/party/PartyDiagram.tsx": 227,
+    "apps/web/src/surfaces/party/PartyPage.test.tsx": 144,
+    "apps/web/src/surfaces/party/PartyPage.tsx": 116,
+    "apps/web/src/surfaces/party/PartyStage.tsx": 194,
+    "apps/web/src/surfaces/party/QrCode.tsx": 41,
+    "apps/web/src/surfaces/party/roomClient.ts": 137,
+    "apps/web/src/surfaces/party/voicePhrases.test.ts": 19,
+    "apps/web/src/surfaces/party/voicePhrases.ts": 26,
     "apps/web/src/telemetry.test.ts": 285,
     "apps/web/src/telemetry.ts": 273,
     "apps/web/src/TelemetryCharts.test.tsx": 197,
